remove image dep and rename misc.py -> helpers.py
