+++ ClemBot.Bot/bot/cogs/moderation_cog/infractions_cog.py
@@ -1,7 +1,7 @@
 import logging
 import typing as t
 from datetime import datetime
-from bot.utils.misc import chunk_sequence
+from bot.utils.helpers import chunk_sequence
 
 import discord
 import discord.ext.commands as commands
