+++ ClemBot.Bot/bot/utils/helpers.py
